+++ includes/bricks-elements/loader.php
@@ -99,7 +99,7 @@ function evk_elements_registry(): array {
             'name'  => 'evk-circular-menu',
             'file'  => $dir . 'evoke-circular-menu/element.php',
             'consts'=> [
-                'EVK_CIRCULAR_MENU_VERSION' => '1.3.0',
+                'EVK_CIRCULAR_MENU_VERSION' => '1.3.1',
                 'EVK_CIRCULAR_MENU_URL'     => $url . 'evoke-circular-menu/',
                 'EVK_CIRCULAR_MENU_PATH'    => $dir . 'evoke-circular-menu/',
             ],
@@ -114,7 +114,7 @@ function evk_elements_registry(): array {
             'name'  => 'evk-offcanvas-menu',
             'file'  => $dir . 'evoke-offcanvas-menu/element.php',
             'consts'=> [
-                'EVK_OFFCANVAS_MENU_VERSION' => '1.7.0',
+                'EVK_OFFCANVAS_MENU_VERSION' => '1.7.1',
                 'EVK_OFFCANVAS_MENU_URL'     => $url . 'evoke-offcanvas-menu/',
                 'EVK_OFFCANVAS_MENU_PATH'    => $dir . 'evoke-offcanvas-menu/',
             ],